+++ src/main.js
@@ -1,4 +1,3 @@
-import {generatePoint} from './mock/point.js';
 import InfoPresenter from "./presenter/info.js";
 import TripPresenter from "./presenter/trip.js";
 import FilterPresenter from "./presenter/filter.js";
@@ -14,20 +13,7 @@ const END_POINT = `https://12.ecmascript.pages.academy/big-trip`;
 
 const api = new Api(END_POINT, AUTHORIZATION);
 
-api.getPoints().then((points) => {
-  console.log(points);
-  // Есть проблема: cтруктура объекта похожа, но некоторые ключи называются иначе,
-  // а ещё на сервере используется snake_case, а у нас camelCase.
-  // Можно, конечно, переписать часть нашего клиентского приложения, но зачем?
-  // Есть вариант получше - паттерн "Адаптер"
-});
-
-const POINT_COUNT = 20;
-const points = new Array(POINT_COUNT).fill().map(generatePoint);
-console.log(points);
-
 const pointsModel = new PointsModel();
-pointsModel.setPoints(points);
 
 const sitePageBodyContainer = document.querySelector(`.page-body__page-main .page-body__container`);
 
@@ -79,8 +65,15 @@ const handleMenuClick = (menuItem) => {
 };
 // Info
 const infoPresenter = new InfoPresenter(mainElement, pointsModel, handleMenuClick);
-infoPresenter.init();
+
 
 // Trip
 const tripPresenter = new TripPresenter(pointsModel, filterModel);
-tripPresenter.init();
+
+
+api.getPoints().then((points) => {
+  pointsModel.setPoints(points);
+  console.log(points);
+  infoPresenter.init();
+  tripPresenter.init();
+});
