+++ src/view/point.js
@@ -9,7 +9,7 @@ const createTypeTemplate = (type, city) => {
   return `<div class="event__type">
             <img class="event__type-icon" width="42" height="42" src="img/icons/${type.toLowerCase()}.png" alt="Event type icon">
           </div>
-          <h3 class="event__title">${type} to ${he.encode(city)}</h3>`;
+          <h3 class="event__title">${type[0].toUpperCase() + type.slice(1)} to ${he.encode(city)}</h3>`;
 };
 
 // Шаблон для даты
@@ -39,7 +39,7 @@ const createOfferTemplate = (offers) => {
           ${Object.values(offers).map((element) => `<li class="event__offer">
               <span class="event__offer-title">${element.title}</span>
               &plus;
-              &euro;&nbsp;<span class="event__offer-price">${element.cost}</span>
+              &euro;&nbsp;<span class="event__offer-price">${element.price}</span>
             </li>`).slice(-2).join(``)}
 
           </ul>
